Eliminate stack arrays.
The only real difference between iARRAY and iREFARRAY is that local
iARRAYs are allocated on the stack, whereas iREFARRAY declarations are
allocated on the heap (with GENARRAY).

Long-term, all arrays should be allocated in the VM, not the compiler.
Let's bite the bullet now and remove this oddity. This will make it
simpler to unify how arrays work.

With this change, the stack exclusively contains scalars. All local
arrays are allocated on the heap, and their addresses are stored in a
stack variable.

Global and static arrays are unaffected since all global arrays are
iARRAYs.