Remove iARRAY.
This removes iARRAY as a storage class and makes arrays an actual
built-in type. While this may seem boring, and has very little practical
effect right now, it is the biggest change to the compiler we've done to
date (including the refactoring to move to an AST). It fundamentally
changes type checking, begins removing matchtag(), and begins removing
Old Syntax.

Because arrays now fully participate in the type system, our type checks
become much more comprehensive, and we can start building out unified
type checking logic rather than hard-coding one-off checks all over the
codebase. It's more work, but it increases the rigor of the type system,
and it allows us to represent things we couldn't before.

For example, we can now have an iEXPRESSION that returns an array.
Things like this are now possible:

    char c = "hello"[1];

This was not possible before, because expressions didn't have a way to
attach array information other than via having a backing variable. That
also means we can eliminate hacks like creating temporary Decls to
represent intermediate values (like we used to in FieldAccessExpr).

As a consequence of now having a proper type checker, it is time to
start removing "old" syntax - the pre-transitional, tag-based syntax. We
will slowly begin removing matchtag() and turning type warnings into
errors.

Type-related error messages have also been improved:

    error 047: array sizes do not match, or destination array is too small

Becomes:

    error 450: no viable conversion from "float[4]" to "float[3]"

In this change we are also removing support for tagged functions that
return arrays. The following:

    String:blah() {
        char x[10] = "hello";
        return x;
    }

Is no longer valid. The return type must use typed syntax:

    char[] blah() {
        char x[10] = "hello";
        return x;
    }

Similarly, forgetting the brackets is also an error.

As an added bonus, this change has a 10% reduction in compiler memory
usage, due to removing lots of unnecessary information attached to the
AST.

Finally, there are some bugfixes here to error reporting, which were
discovered when running the corpus against the new type checker.