Only Ignore Unused Stocks when they are Included
This changes to also check swhether the VarDecl is from the main source file or not.
Checking this seems easy enough with the new source file tracking.

If a variable is from the main source file even if it is marked with stock we shall now report a warning.

We could also ignore any unused variable that is not from the main file.
But I shall stick with only stock for now.

Testing

Extended out the test to cover both the pass and fail cases.