Allow natives to return structs and arrays.
Natives can now return enum structs and fixed-size arrays. The address
is passed as a hidden first parameter. For example:

    native Struct DoStuff(int x, int y);

Will have:

    params[0] = 3
    params[1] = the address of the output struct
    params[2] = x
    params[3] = y

The hidden address points to a fully initialized array or enum struct.
In the case of a multi-dimensional array, the indirection vectors are
already placed. All values/members are initialized to zero.

The native MUST return params[1] on success, otherwise, behavior is
undefined.

It is recommended that when defining C++ definitions of SourcePawn
structs, to use `#pragma pack(push, 1)` and `#pragma pack(pop)`. In
addition, the only types that should appear are "cell_t" and "float". No
other type should be used for scalars. For arrays, "char" may be used,
but not as a scalar.

When using char arrays, the C++ definition must be careful to convert
the char size to a cell-aligned size. For example:

    enum struct MyStruct {
        int x;
        int y;
        char message[50];
    }

Should look like this in C++:

    #pragma pack(push, 1)
    struct MyStruct {
        cell_t x;
        cell_t y;
        char message[CharArraySize<50>::bytes];
    };
    #pragma pack(pop)